requirements.txt
asyncpg==0.30.0
coinbase_advanced_py==1.8.1
python-dotenv==1.0.1
Flask==3.1.0
Flask_Cors==5.0.0
pandas==2.2.3
plotly==5.24.1
psycopg2==2.9.10
Requests==2.32.3
urllib3==2.2.3
watchdog==6.0.0
websockets==13.1
Werkzeug==3.0.6
